// src/lib/embody/lib/validating/types.ts
type SourcePosition = {
	readonly line: number;
	readonly column: number;
};
// ...
type SourceRange = {
	readonly start: SourcePosition;
	readonly end: SourcePosition;
};